Remove bookmarks and details card, comment out tests for unfinished components.
